+++ src/library/components/projects/projectItem/Info.js
@@ -3,7 +3,7 @@ import DevIcons from "./DevIcons"
 import Title from "./Title"
 import About from "./About"
 import DeployedButtons from "./DeployedButtons"
-import { Grid, Paper, Button, Typography } from "@mui/material"
+import { Grid, Paper } from "@mui/material"
 import { useTheme } from "@mui/material/styles"
 
 import Fade from "react-reveal/Fade"
@@ -16,25 +16,14 @@ export default function Info({ project }) {
         <Fade bottom>
             <Paper
                 sx={{
-                    padding: theme.spacing(4),
+                    padding: theme.spacing(3),
                     textAlign: "center",
                     color: theme.palette.text.secondary,
                 }}
             >
                 <Title project={project} />
                 <About project={project} />
-
-                <Grid
-                    item
-                    container
-                    direction="row"
-                    justifyContent="space-around"
-                    alignItems="center"
-                    sx={{ height: "max-content", width: "100%" }}
-                >
-                    <DevIcons iconList={stack} />
-                </Grid>
-
+                <DevIcons iconList={stack} />
                 <DeployedButtons project={project} />
             </Paper>
         </Fade>
